devenv: split out cli
