+++ devenv/src/main.rs
@@ -1,8 +1,10 @@
+mod cli;
 mod command;
 mod config;
 mod log;
 
-use clap::{crate_version, Parser, Subcommand};
+use clap::{crate_version, Parser};
+use cli::{Cli, Commands, ContainerCommand, InputsCommand, ProcessesCommand};
 use cli_table::{print_stderr, Table, WithTitle};
 use include_dir::{include_dir, Dir};
 use miette::{bail, IntoDiagnostic, Result, WrapErr};
@@ -26,240 +28,6 @@ const EXISTING_REQUIRED_FILES: [&str; 1] = [".gitignore"];
 const PROJECT_DIR: Dir = include_dir!("$CARGO_MANIFEST_DIR/init");
 // project vars
 const DEVENV_FLAKE: &str = ".devenv.flake.nix";
-
-#[derive(Parser)]
-#[command(
-    color = clap::ColorChoice::Auto,
-    // for --clean to work with subcommands
-    subcommand_precedence_over_arg = true,
-    dont_delimit_trailing_values = true,
-    about = format!("https://devenv.sh {}: Fast, Declarative, Reproducible, and Composable Developer Environments", crate_version!())
-)]
-struct Cli {
-    #[arg(short, long, global = true, help = "Enable debug log level.")]
-    verbose: bool,
-
-    #[arg(short = 'j', long, global = true, help = "Maximum number of Nix builds at any time.", default_value_t = max_jobs())]
-    max_jobs: u8,
-
-    #[arg(
-        short = 'u',
-        long,
-        help = "Maximum number CPU cores being used by a single build..",
-        default_value = "2"
-    )]
-    cores: u8,
-
-    #[arg(short, long, global = true, default_value_t = default_system())]
-    system: String,
-
-    #[arg(
-        short,
-        long,
-        global = true,
-        help = "Relax the hermeticity of the environment."
-    )]
-    impure: bool,
-
-    #[arg(
-        long,
-        global = true,
-        help = "Disable substituters and consider all previously downloaded files up-to-date."
-    )]
-    offline: bool,
-
-    // TODO: --no-clean?
-    #[arg(
-        short,
-        long,
-        global = true,
-        num_args = 0..,
-        value_delimiter = ',',
-        help = "Ignore existing environment variables when entering the shell. Pass a list of comma-separated environment variables to let through."
-    )]
-    clean: Option<Vec<String>>,
-
-    #[arg(long, global = true, help = "Enter the Nix debugger on failure.")]
-    nix_debugger: bool,
-
-    #[arg(
-        short,
-        long,
-        global = true,
-        num_args = 2,
-        value_delimiter = ' ',
-        help = "Pass additional options to nix commands, see `man nix.conf` for full list."
-    )]
-    nix_option: Vec<String>,
-
-    #[arg(
-        short,
-        long,
-        global = true,
-        num_args = 2,
-        value_delimiter = ' ',
-        help = "Override inputs in devenv.yaml."
-    )]
-    override_input: Vec<String>,
-
-    #[command(subcommand)]
-    command: Commands,
-}
-
-#[derive(Subcommand, Clone)]
-enum Commands {
-    #[command(about = "Scaffold devenv.yaml, devenv.nix, .gitignore and .envrc.")]
-    Init {
-        target: Option<PathBuf>,
-    },
-
-    #[command(about = "Activate the developer environment. https://devenv.sh/basics/")]
-    Shell {
-        cmd: Option<String>,
-        args: Vec<String>,
-    },
-
-    #[command(about = "Update devenv.lock from devenv.yaml inputs. http://devenv.sh/inputs/")]
-    Update {
-        name: Option<String>,
-    },
-
-    #[command(
-        about = "Search for packages and options in nixpkgs. https://devenv.sh/packages/#searching-for-a-file"
-    )]
-    Search {
-        name: String,
-    },
-
-    #[command(
-        alias = "show",
-        about = "Print information about this developer environment."
-    )]
-    Info {},
-
-    #[command(about = "Start processes in the foreground. https://devenv.sh/processes/")]
-    Up {
-        #[arg(help = "Start a specific process.")]
-        process: Option<String>,
-
-        #[arg(short, long, help = "Start processes in the background.")]
-        detach: bool,
-    },
-
-    Processes {
-        #[command(subcommand)]
-        command: ProcessesCommand,
-    },
-
-    #[command(about = "Run tests. http://devenv.sh/tests/", alias = "ci")]
-    Test {
-        #[arg(short, long, help = "Don't override .devenv to a temporary directory.")]
-        dont_override_dotfile: bool,
-    },
-
-    Container {
-        #[arg(short, long)]
-        registry: Option<String>,
-
-        #[arg(long, hide = true)]
-        copy: bool,
-
-        #[arg(long, hide = true)]
-        docker_run: bool,
-
-        #[arg(long)]
-        copy_args: Vec<String>,
-
-        #[arg(hide = true)]
-        name: Option<String>,
-
-        #[command(subcommand)]
-        command: Option<ContainerCommand>,
-    },
-
-    Inputs {
-        #[command(subcommand)]
-        command: InputsCommand,
-    },
-
-    Repl {},
-
-    #[command(
-        about = "Deletes previous shell generations. See http://devenv.sh/garbage-collection"
-    )]
-    Gc {},
-
-    #[command(about = "Build any attribute in devenv.nix.")]
-    Build {
-        #[arg(num_args=1..)]
-        attributes: Vec<String>,
-    },
-
-    #[command(about = "Print the version of devenv.")]
-    Version {},
-
-    #[clap(hide = true)]
-    Assemble,
-
-    #[clap(hide = true)]
-    PrintDevEnv {
-        #[arg(long)]
-        json: bool,
-    },
-
-    #[clap(hide = true)]
-    GenerateJSONSchema,
-}
-
-#[derive(Subcommand, Clone)]
-#[clap(about = "Start or stop processes. https://devenv.sh/processes/")]
-enum ProcessesCommand {
-    #[command(alias = "start", about = "Start processes in the foreground.")]
-    Up {
-        process: Option<String>,
-
-        #[arg(short, long, help = "Start processes in the background.")]
-        detach: bool,
-    },
-
-    #[command(alias = "stop", about = "Stop processes running in the background.")]
-    Down {},
-    // TODO: Status/Attach
-}
-
-#[derive(Subcommand, Clone)]
-#[clap(
-    about = "Build, copy, or run a container. https://devenv.sh/containers/",
-    arg_required_else_help(true)
-)]
-enum ContainerCommand {
-    #[command(about = "Build a container.")]
-    Build { name: String },
-
-    #[command(about = "Copy a container to registry.")]
-    Copy { name: String },
-
-    #[command(about = "Run a container.")]
-    Run { name: String },
-}
-
-#[derive(Subcommand, Clone)]
-#[clap(about = "Add an input to devenv.yaml. https://devenv.sh/inputs/")]
-enum InputsCommand {
-    #[command(about = "Add an input to devenv.yaml.")]
-    Add {
-        #[arg(help = "The name of the input.")]
-        name: String,
-
-        #[arg(
-            help = "See https://devenv.sh/reference/yaml-options/#inputsnameurl for possible values."
-        )]
-        url: String,
-
-        #[arg(short, long, help = "What inputs should follow your inputs?")]
-        follows: Vec<String>,
-    },
-}
 
 struct App {
     cli: Cli,
@@ -1230,27 +998,6 @@ fn symlink_force(logger: &log::Logger, link_path: &Path, target: &Path) {
     });
 }
 
-fn default_system() -> String {
-    let arch = if cfg!(target_arch = "aarch64") {
-        "aarch64"
-    } else if cfg!(target_arch = "x86_64") {
-        "x86_64"
-    } else {
-        "unknown architecture"
-    };
-
-    let os = if cfg!(target_os = "linux") {
-        "linux"
-    } else if cfg!(target_os = "windows") {
-        "windows"
-    } else if cfg!(target_os = "macos") {
-        "darwin" // macOS is referred to as "darwin" in target triples
-    } else {
-        "unknown OS"
-    };
-    format!("{arch}-{os}")
-}
-
 fn get_now_with_nanoseconds() -> String {
     let now = SystemTime::now();
     let duration = now.duration_since(UNIX_EPOCH).expect("Time went backwards");
@@ -1323,23 +1070,4 @@ fn cleanup_symlinks(root: &Path) -> (Vec<PathBuf>, Vec<PathBuf>) {
     }
 
     (to_gc, removed_symlinks)
-}
-
-fn max_jobs() -> u8 {
-    let num_cpus = std::thread::available_parallelism().unwrap_or_else(|e| {
-        eprintln!("Failed to get number of logical CPUs: {}", e);
-        std::num::NonZeroUsize::new(1).unwrap()
-    });
-    (num_cpus.get() / 2).try_into().unwrap()
-}
-
-#[cfg(test)]
-mod tests {
-    use super::Cli;
-
-    #[test]
-    fn verify_cli() {
-        use clap::CommandFactory;
-        Cli::command().debug_assert()
-    }
 }
